+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   `DUNE_ENABLE_PYTHONBINDINGS` needs to be turned on (default is off).
   Furthermore, flags for either shared library or position independent code
   needs to be used.
+- Properly implement the `canCommunicate` capability for `UGGrid` and `IdentityGrid`.
 
 ## Deprecations and removals